requirements.txt
PyYAML>=6.0.0
pytest>=8.2.0
pytest-cov>=5.0.0
